+++ kos/amqp_consumer/cross-build.sh
@@ -1,21 +1,23 @@
 #!/usr/bin/env bash
 
-# © 2024 AO Kaspersky Lab
+# © 2025 AO Kaspersky Lab
 # Licensed under the MIT License
+
+set -e
 
 usage() {
     echo "Usage: $0 <TARGET> [SDK_PATH] "
     echo ""
     echo "Available targets":
-    echo "  qemu     - build for QEMU."
-    echo "  rpi      - build for Raspberry Pi 4 B."
+    echo "  qemu - build for QEMU."
+    echo "  hw   - build for hardware."
     echo ""
-    echo "Optional arguments:"
+    echo "Optional argument:"
     echo "  SDK_PATH - Path to the KasperskyOS SDK."
     echo ""
     echo "Examples:"
     echo "  ./cross-build.sh qemu"
-    echo "  ./cross-build.sh rpi \"/opt/KasperskyOS-Community-Edition-<version>\""
+    echo "  ./cross-build.sh hw \"/opt/KasperskyOS-Community-Edition-<version>\""
     echo ""
 }
 
@@ -29,12 +31,12 @@ fi
 
 if [ "$1" == "qemu" ]
 then
-    TARGET_PLATFORM="sim"
-elif [ "$1" == "rpi" ]
+    TARGET="sim"
+elif [ "$1" == "hw" ]
 then
-    TARGET_PLATFORM="sd-image"
+    TARGET="sd-image"
 else
-    echo "Unknown target platform is specified."
+    echo "Unknown target is specified."
     usage
     exit 1
 fi
@@ -48,7 +50,7 @@ SCRIPT_DIR="$(dirname "$(realpath "$0")")"
 BUILD="${SCRIPT_DIR}/build"
 
 export LANG=C
-export TARGET="aarch64-kos"
+export TARGET_PLATFORM="aarch64-kos"
 export PKG_CONFIG=""
 export INSTALL_PREFIX="$BUILD/../install"
 
@@ -76,11 +78,15 @@ if [ "$BUILD_WITH_GCC" == "y" ];then
     TOOLCHAIN_SUFFIX="-gcc"
 fi
 
+(cd $SCRIPT_DIR/../../ && git submodule update --init --depth=1 third_party/openssl && cd third_party/openssl && ./config && make build_generated)
+
 "$SDK_PREFIX/toolchain/bin/cmake" -G "Unix Makefiles" -B "$BUILD" \
       -D CMAKE_BUILD_TYPE:STRING=Debug \
       -D CMAKE_INSTALL_PREFIX:STRING="$INSTALL_PREFIX" \
-      -D BOARD:STRING="RPI4_BCM2711" \
       -D CMAKE_EXTERNAL_PREFIX:STRING="$PWD/../.." \
-      -D CMAKE_FIND_ROOT_PATH="$([[ -f "$SCRIPT_DIR/additional_cmake_find_root.txt" ]] && cat "$SCRIPT_DIR/additional_cmake_find_root.txt")$PREFIX_DIR/sysroot-$TARGET" \
-      -D CMAKE_TOOLCHAIN_FILE="$SDK_PREFIX/toolchain/share/toolchain-$TARGET$TOOLCHAIN_SUFFIX.cmake" \
-      "$SCRIPT_DIR/" && "$SDK_PREFIX/toolchain/bin/cmake" --build "$BUILD" --target $TARGET_PLATFORM
+      -D CMAKE_FIND_ROOT_PATH="$([[ -f "$SCRIPT_DIR/additional_cmake_find_root.txt" ]] && cat "$SCRIPT_DIR/additional_cmake_find_root.txt")$PREFIX_DIR/sysroot-$TARGET_PLATFORM" \
+      -D CMAKE_TOOLCHAIN_FILE="$SDK_PREFIX/toolchain/share/toolchain-$TARGET_PLATFORM$TOOLCHAIN_SUFFIX.cmake" \
+      -D SDK_PREFIX=$SDK_PREFIX \
+      -D IMAGE=$TARGET \
+      -D OPENSSL_INCLUDE_DIR="$SCRIPT_DIR/../../third_party/openssl/include/" \
+      "$SCRIPT_DIR/" && "$SDK_PREFIX/toolchain/bin/cmake" --build "$BUILD" --target $TARGET