+++ ChangeLog.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## Release for KasperskyOS Community Edition SDK 1.3 - 2025-07
+## Changed
+- Replaced deprecated word "entity" by "task class"
+- Fixed ntpd process connections
+- Updated STANDARD_GNU compiler flag to version 17
+- Added support for the Radxa ROCK 3A hardware platform
+- Added OpenSSL git submodule
+
 ## Release for KasperskyOS Community Edition SDK 1.2 - 2024-06-26
 - Updated documentation
 - Added shared libs support